V0.1.5.0
V0.1.5.0
Added check to part startup for fields
stockOverlayDepthMaskShaderTransform &
transparentPodDepthMaskShaderTransform if they cannot be found
then processing is sped up (effiency change).
Added bool field combineDepthMaskShaders. If this is set to True
(Default is false) then the transparentPodDepthMaskShaderTransform is
treated as an ADD_ON to
the stockOverlayDepthMaskShaderTransform mesh. and BOTH will be turned
on in JSIATP mode and the transparentPodDepthMaskShaderTransform will be
turned off and the stockOverlayDepthMaskShaderTransform will be turned
ON when in Stock overlay mode. This is to support different model
configurations depending on what the modeller/modder requires. See WIKI
for more details (added pictures to the WIKI as well).